move functionality that can be shared outside of ElasticsearchNetIntegration
